+++ packages/create-app/common/_npmrc
@@ -1,2 +1,3 @@
 ## force pnpm to hoist
-shamefully-hoist = true
+shamefully-hoist=true
+strict-peer-dependencies=false
