+++ packages/create-app/index.js
@@ -15,6 +15,7 @@ import { getUser } from './lib/index.js';
 import { checkServerInstalled, checkServerRunning, checkSatisfiedVersion, getServerDirectory } from './lib/server.js';
 import { toBlockletDid } from './lib/did.js';
 import { initGitRepo } from './lib/git.js';
+import { copy, emptyDir, isEmpty, isValidPackageName, toValidPackageName } from './lib/utils.js';
 
 const { yellow, red, green, cyan, blue, bold, magenta } = chalk;
 
@@ -516,58 +517,6 @@ async function init() {
     // else {
     //   console.warn(`\n${yellow('No .env file found, please add one.')}`);
     // }
-  }
-}
-
-// common functions
-function copy(src, dest) {
-  const stat = fs.statSync(src);
-  if (stat.isDirectory()) {
-    copyDir(src, dest);
-  } else {
-    fs.copyFileSync(src, dest);
-  }
-}
-
-function isValidPackageName(projectName) {
-  return /^(?:@[a-z0-9-*~][a-z0-9-*._~]*\/)?[a-z0-9-~][a-z0-9-._~]*$/.test(projectName);
-}
-
-function toValidPackageName(projectName) {
-  return projectName
-    .trim()
-    .toLowerCase()
-    .replace(/\s+/g, '-')
-    .replace(/^[._]/, '')
-    .replace(/[^a-z0-9-~]+/g, '-');
-}
-
-function copyDir(srcDir, destDir) {
-  fs.mkdirSync(destDir, { recursive: true });
-  for (const file of fs.readdirSync(srcDir)) {
-    const srcFile = path.resolve(srcDir, file);
-    const destFile = path.resolve(destDir, file);
-    copy(srcFile, destFile);
-  }
-}
-
-function isEmpty(_path) {
-  return fs.readdirSync(_path).length === 0;
-}
-
-function emptyDir(dir) {
-  if (!fs.existsSync(dir)) {
-    return;
-  }
-  for (const file of fs.readdirSync(dir)) {
-    const abs = path.resolve(dir, file);
-    // baseline is Node 12 so can't use rmSync :(
-    if (fs.lstatSync(abs).isDirectory()) {
-      emptyDir(abs);
-      fs.rmdirSync(abs);
-    } else {
-      fs.unlinkSync(abs);
-    }
   }
 }
