+++ packages/create-app/package.json
@@ -29,9 +29,9 @@
     "test:run": "vitest run"
   },
   "dependencies": {
-    "@arcblock/did": "^1.17.5",
-    "@ocap/mcrypto": "^1.17.5",
-    "@ocap/util": "^1.17.5",
+    "@arcblock/did": "^1.17.6",
+    "@ocap/mcrypto": "^1.17.6",
+    "@ocap/util": "^1.17.6",
     "boxen": "^6.2.1",
     "ejs": "^3.1.8",
     "envfile": "^6.17.0",
@@ -47,8 +47,8 @@
   },
   "devDependencies": {
     "@arcblock/eslint-config-base": "0.2.2",
-    "eslint": "^8.19.0",
+    "eslint": "^8.20.0",
     "prettier": "^2.7.1",
-    "vitest": "^0.18.0"
+    "vitest": "^0.19.1"
   }
 }
