+++ packages/create-app/templates/blocklet-page-static/.blocklet-page/blocks/CustomBlock.mdx
