+++ packages/create-app/templates/blocklet-page-static/.blocklet-page/composes/MyLayout.mdx
