+++ packages/create-app/templates/blocklet-page-static/.blocklet-page/config.yml
