+++ packages/create-app/templates/blocklet-page-static/package.json
@@ -10,6 +10,6 @@
     "upload": "npm run bundle && blocklet upload .blocklet/release/blocklet.json"
   },
   "dependencies": {
-    "@xmark/cli": "^2.0.64"
+    "@xmark/cli": "^2.2.3"
   }
 }
