+++ packages/create-app/templates/blocklet-page-static/page/index.md
