+++ packages/create-app/templates/express-api/package.json
@@ -16,8 +16,8 @@
   "license": "ISC",
   "dependencies": {
     "@arcblock/did-auth-storage-nedb": "^1.6.3",
-    "@blocklet/sdk": "^1.8.2",
-    "@ocap/client": "^1.17.4",
+    "@blocklet/sdk": "^1.8.4",
+    "@ocap/client": "^1.17.6",
     "dotenv-flow": "^3.2.0",
     "express": "^4.18.1"
   },
