+++ packages/create-app/templates/nextjs-dapp/package.json
@@ -25,21 +25,21 @@
     ]
   },
   "dependencies": {
-    "@arcblock/did-auth": "^1.17.4",
+    "@arcblock/did-auth": "^1.17.6",
     "@arcblock/did-auth-storage-nedb": "^1.6.3",
-    "@blocklet/sdk": "^1.8.2",
-    "@ocap/client": "^1.17.4",
-    "@ocap/mcrypto": "^1.17.4",
-    "@ocap/wallet": "^1.17.4",
+    "@blocklet/sdk": "^1.8.4",
+    "@ocap/client": "^1.17.6",
+    "@ocap/mcrypto": "^1.17.6",
+    "@ocap/wallet": "^1.17.6",
     "dotenv-flow": "^3.2.0",
     "express": "^4.18.1",
-    "next": "12.1.6",
-    "react": "18.1.0",
-    "react-dom": "18.1.0"
+    "next": "12.2.3",
+    "react": "18.2.0",
+    "react-dom": "18.2.0"
   },
   "devDependencies": {
-    "eslint": "8.16.0",
-    "eslint-config-next": "12.1.6",
+    "eslint": "8.20.0",
+    "eslint-config-next": "12.2.3",
     "husky": "^8.0.1",
     "lint-staged": "^12.5.0",
     "nodemon": "^2.0.19",
