+++ packages/create-app/templates/react-gun-dapp/package.json
@@ -39,12 +39,12 @@
     ]
   },
   "dependencies": {
-    "@arcblock/did-auth": "^1.17.4",
+    "@arcblock/did-auth": "^1.17.6",
     "@arcblock/did-auth-storage-nedb": "^1.6.3",
-    "@blocklet/sdk": "^1.8.2",
-    "@ocap/client": "^1.17.4",
-    "@ocap/mcrypto": "^1.17.4",
-    "@ocap/wallet": "^1.17.4",
+    "@blocklet/sdk": "^1.8.4",
+    "@ocap/client": "^1.17.6",
+    "@ocap/mcrypto": "^1.17.6",
+    "@ocap/wallet": "^1.17.6",
     "axios": "^0.27.2",
     "compression": "^1.7.4",
     "cookie-parser": "^1.4.6",
@@ -61,7 +61,7 @@
   "devDependencies": {
     "@arcblock/eslint-config": "^0.2.2",
     "@craco/craco": "^6.4.5",
-    "eslint": "^8.19.0",
+    "eslint": "^8.20.0",
     "husky": "^8.0.1",
     "lint-staged": "^12.5.0",
     "nodemon": "^2.0.19",
