+++ packages/create-app/templates/react-static/package.json
@@ -42,7 +42,7 @@
   "devDependencies": {
     "@arcblock/eslint-config": "^0.2.2",
     "@craco/craco": "^6.4.5",
-    "eslint": "^8.19.0",
+    "eslint": "^8.20.0",
     "husky": "^8.0.1",
     "lint-staged": "^12.5.0",
     "prettier": "^2.7.1",
