+++ packages/create-app/templates/solidjs-dapp/package.json
@@ -20,27 +20,27 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "eslint": "^8.19.0",
+    "eslint": "^8.20.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-solid": "0.6.0",
+    "eslint-plugin-solid": "0.7.1",
     "husky": "^8.0.1",
     "lint-staged": "^12.5.0",
     "nodemon": "^2.0.19",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.7.1",
-    "vite": "^2.9.14",
+    "vite": "^3.0.3",
     "vite-plugin-html": "^3.2.0",
-    "vite-plugin-solid": "^2.2.6",
+    "vite-plugin-solid": "^2.3.0",
     "vite-plugin-blocklet": "^0.4.26"
   },
   "dependencies": {
-    "@arcblock/did-auth": "^1.17.4",
+    "@arcblock/did-auth": "^1.17.6",
     "@arcblock/did-auth-storage-nedb": "^1.6.3",
-    "@blocklet/sdk": "^1.8.2",
-    "@ocap/client": "^1.17.4",
-    "@ocap/mcrypto": "^1.17.4",
-    "@ocap/wallet": "^1.17.4",
+    "@blocklet/sdk": "^1.8.4",
+    "@ocap/client": "^1.17.6",
+    "@ocap/mcrypto": "^1.17.6",
+    "@ocap/wallet": "^1.17.6",
     "axios": "^0.27.2",
     "compression": "^1.7.4",
     "cookie-parser": "^1.4.6",
@@ -49,7 +49,7 @@
     "express": "^4.18.1",
     "express-async-errors": "^3.1.1",
     "express-history-api-fallback": "^2.2.1",
-    "solid-js": "^1.4.7"
+    "solid-js": "^1.4.8"
   },
   "lint-staged": {
     "*.{mjs,js,vue}": [
