+++ packages/create-app/templates/solidjs-static/package.json
@@ -16,20 +16,20 @@
   },
   "license": "MIT",
   "devDependencies": {
-    "eslint": "^8.19.0",
+    "eslint": "^8.20.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-solid": "0.6.0",
+    "eslint-plugin-solid": "0.7.1",
     "husky": "^8.0.1",
     "lint-staged": "^12.5.0",
     "prettier": "^2.7.1",
-    "vite": "^2.9.14",
+    "vite": "^3.0.3",
     "vite-plugin-html": "^3.2.0",
-    "vite-plugin-solid": "^2.2.6",
+    "vite-plugin-solid": "^2.3.0",
     "vite-plugin-blocklet": "^0.4.26"
   },
   "dependencies": {
-    "solid-js": "^1.4.7"
+    "solid-js": "^1.4.8"
   },
   "lint-staged": {
     "*.{mjs,js,vue}": [
