+++ packages/create-app/templates/svelte-static/package.json
@@ -15,8 +15,8 @@
     "prepare": "husky install"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "1.0.0-next.47",
-    "eslint": "^8.19.0",
+    "@sveltejs/vite-plugin-svelte": "1.0.1",
+    "eslint": "^8.20.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.2.1",
     "eslint-plugin-svelte3": "^4.0.0",
@@ -24,7 +24,7 @@
     "lint-staged": "^12.5.0",
     "prettier": "^2.7.1",
     "svelte": "^3.49.0",
-    "vite": "^2.9.14",
+    "vite": "^3.0.3",
     "vite-plugin-html": "^3.2.0",
     "vite-plugin-blocklet": "^0.4.26"
   },
