+++ packages/create-app/templates/vue-dapp/package.json
@@ -18,12 +18,12 @@
     "prepare": "husky install"
   },
   "dependencies": {
-    "@arcblock/did-auth": "^1.17.4",
+    "@arcblock/did-auth": "^1.17.6",
     "@arcblock/did-auth-storage-nedb": "^1.6.3",
-    "@blocklet/sdk": "^1.8.2",
-    "@ocap/client": "^1.17.4",
-    "@ocap/mcrypto": "^1.17.4",
-    "@ocap/wallet": "^1.17.4",
+    "@blocklet/sdk": "^1.8.4",
+    "@ocap/client": "^1.17.6",
+    "@ocap/mcrypto": "^1.17.6",
+    "@ocap/wallet": "^1.17.6",
     "axios": "^0.27.2",
     "compression": "^1.7.4",
     "cookie-parser": "^1.4.6",
@@ -36,17 +36,17 @@
     "vue": "^3.2.37"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue": "^2.3.3",
-    "eslint": "^8.19.0",
+    "@vitejs/plugin-vue": "^3.0.1",
+    "eslint": "^8.20.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-vue": "^9.2.0",
+    "eslint-plugin-vue": "^9.3.0",
     "husky": "^8.0.1",
     "lint-staged": "^12.5.0",
     "nodemon": "^2.0.19",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.7.1",
-    "vite": "^2.9.14",
+    "vite": "^3.0.3",
     "vite-plugin-blocklet": "^0.4.26"
   },
   "lint-staged": {
