+++ packages/create-app/templates/vue-static/package.json
@@ -17,15 +17,15 @@
     "vue": "^3.2.37"
   },
   "devDependencies": {
-    "@vitejs/plugin-vue": "^2.3.3",
-    "eslint": "^8.19.0",
+    "@vitejs/plugin-vue": "^3.0.1",
+    "eslint": "^8.20.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-vue": "^9.2.0",
+    "eslint-plugin-vue": "^9.3.0",
     "husky": "^8.0.1",
     "lint-staged": "^12.5.0",
     "prettier": "^2.7.1",
-    "vite": "^2.9.14",
+    "vite": "^3.0.3",
     "vite-plugin-html": "^3.2.0",
     "vite-plugin-blocklet": "^0.4.26"
   },
