+++ packages/docs-site/package.json
@@ -7,15 +7,15 @@
     "start": "xmark start",
     "build": "xmark build",
     "clean": "rm -rf .blocklet",
-    "bundle": "npm run clean && npm run build -- --base=/.blocklet/proxy/z8iZxR98sTfBLymQKPJYE85a95AghSCkLd3bb/ && blocklet bundle --create-release",
+    "bundle": "npm run clean && npm run build && blocklet bundle --create-release",
     "deploy": "npm run bundle && blocklet deploy .blocklet/bundle",
     "deploy:child": "npm run bundle && blocklet deploy .blocklet/bundle --app-did=z8iZuvjcXn3Lj1DXsaKkAPvL4y1nf5SSsUsXc --mount-point=docs",
     "upload": "npm run bundle && blocklet upload .blocklet/release/blocklet.json"
   },
   "dependencies": {
-    "@xmark/cli": "^2.0.64",
-    "@xmark/client": "^2.0.64",
-    "@xmark/theme-docs": "^2.0.64"
+    "@xmark/cli": "^2.2.3",
+    "@xmark/client": "^2.2.3",
+    "@xmark/theme-docs": "^2.2.3"
   },
   "version": "0.4.26"
 }
