+++ plugins/vite-plugin-wss-hmr/package.json
@@ -20,7 +20,7 @@
   "author": "Liang Yongzhuo",
   "license": "ISC",
   "devDependencies": {
-    "rollup": "^2.76.0",
+    "rollup": "^2.77.1",
     "vite": "^2.9.14"
   }
 }
